major refactor split everything into individual modules
