+++ index.html
@@ -147,7 +147,7 @@
   <!-- Scripts -->
   <script src="https://code.jquery.com/jquery-3.7.1.min.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.bundle.min.js"></script>
-  <script type="module" src="./src/main.js"></script>
+  <script type="module" src="./src/index.js"></script>
 </body>
 
 </html>
